# int blends, some simd, all around faster