… tests/benchmark_blends.nim
… tests/benchmark_images_draw.nim
… tests/images/svg/Ghostscript_Tiger.png