before:

```
draw bmNormal ....................... 6.953 ms    ±0.377  x10
draw Smooth bmNormal ............... 16.723 ms    ±0.142  x10

run_files.nim Driving - Navigation: 3.28490686416626
svg decode ........................ 285.880 ms    ±3.089  x10
```

after:

```
draw bmNormal ....................... 4.625 ms    ±0.319  x10
draw Smooth bmNormal ............... 15.330 ms    ±0.228  x10

run_files.nim Driving - Navigation: 2.821513175964356
svg decode ........................ 235.707 ms    ±1.834  x10
```