@guzba 2020-12-09
```
blendNormal ......................... 5.595 ms    ±0.113  x10
blendNormalFloats .................. 17.934 ms    ±0.071  x10
blendDarken ......................... 7.807 ms    ±0.223  x10
blendDarkenFloats .................. 30.119 ms    ±0.119  x10
blendMultiply ....................... 7.496 ms    ±0.289  x10
blendMultiplyFloats ................ 29.610 ms    ±0.169  x10
blendLinearBurn ..................... 7.839 ms    ±0.239  x10
blendLinearBurnFloats .............. 29.810 ms    ±0.168  x10
blendColorBurn ..................... 14.482 ms    ±0.164  x10
blendColorBurnFloats ............... 27.079 ms    ±0.104  x10
blendLighten ........................ 7.817 ms    ±0.213  x10
blendLightenFloats ................. 29.973 ms    ±0.087  x10
blendScreen ......................... 7.453 ms    ±0.249  x10
blendScreenFloats .................. 30.946 ms    ±0.207  x10
blendLinearDodge .................... 7.420 ms    ±0.225  x10
blendLinearDodgeFloats ............. 29.500 ms    ±0.129  x10
blendColorDodge .................... 14.358 ms    ±0.141  x10
blendColorDodgeFloats .............. 33.459 ms    ±0.886  x10
blendOverlay ........................ 7.781 ms    ±0.403  x10
blendOverlayFloats ................. 28.561 ms    ±0.348  x10
blendSoftLight ...................... 4.042 ms    ±0.058  x10
blendSoftLightFloats ............... 33.186 ms    ±0.373  x10
blendHardLight ...................... 7.488 ms    ±0.256  x10
blendHardLightFloats ............... 27.753 ms    ±0.286  x10
blendDifference ..................... 7.829 ms    ±0.212  x10
blendDifferenceFloats .............. 25.506 ms    ±0.112  x10
blendExclusion ...................... 7.437 ms    ±0.201  x10
blendExclusionFloats ............... 30.530 ms    ±0.168  x10
blendHue ........................... 38.105 ms    ±0.236  x10
blendHueFloats ..................... 38.907 ms    ±0.183  x10
blendSaturation .................... 39.074 ms    ±0.207  x10
blendSaturationFloats .............. 39.099 ms    ±0.178  x10
blendColor ......................... 27.228 ms    ±0.158  x10
blendColorFloats ................... 29.247 ms    ±0.234  x10
blendLuminosity .................... 37.434 ms    ±0.465  x10
blendLuminosityFloats .............. 36.854 ms    ±0.108  x10
blendMask ........................... 1.435 ms    ±0.006  x10
blendMaskFloats ..................... 9.388 ms    ±0.269  x10
blendSubtractMask ................... 1.446 ms    ±0.040  x10
blendSubtractMaskFloats ............. 1.431 ms    ±0.001  x10
blendIntersectMask .................. 1.432 ms    ±0.004  x10
blendIntersectMaskFloats ............ 1.432 ms    ±0.003  x10
blendExcludeMask .................... 1.437 ms    ±0.011  x10
blendExcludeMaskFloats .............. 8.553 ms    ±0.265  x10
```